rename(#12): 코드 컨벤션에 맞춰 디렉토리 명 수정
controller와 service 디렉토리를 제거하고 도메인 디렉토리 내에 위치하도록 수정하였습니다.
