+++ server/src/main/java/com/hexacore/tayo/test/TestController.java
@@ -1,10 +1,9 @@
-package com.hexacore.tayo.test.controller;
+package com.hexacore.tayo.test;
 
 import com.hexacore.tayo.common.DataResponseDto;
 import com.hexacore.tayo.common.PageResponseDto;
 import com.hexacore.tayo.common.ResponseDto;
 import com.hexacore.tayo.test.model.TestDto;
-import com.hexacore.tayo.test.service.TestService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
